all: add utils package, fix order of events when loading config
