+++ internal/test/e2e/cluster.go
@@ -25,8 +25,8 @@ import (
 	"go.sia.tech/renterd/bus"
 	"go.sia.tech/renterd/config"
 	"go.sia.tech/renterd/internal/test"
-	"go.sia.tech/renterd/internal/utils"
 	"go.sia.tech/renterd/stores"
+	"go.sia.tech/renterd/utils"
 	"go.sia.tech/renterd/worker/s3"
 	"go.sia.tech/web/renterd"
 	"go.uber.org/zap"
